Show no search results on empty search for add friends search
