+++ src/components/Friends.js
@@ -3,14 +3,11 @@ import { Column, Container, Row, Center } from './common'
 import { Button } from 'components/Buttons'
 import Image from 'next/image'
 import { Grid } from 'styled-css-grid'
-import {
-  InstantSearch,
-  connectHits,
-  connectSearchBox
-} from 'react-instantsearch-dom'
+import { InstantSearch, connectSearchBox } from 'react-instantsearch-dom'
 import { searchClient } from 'algolia/index'
 import { useState } from 'react'
-import { useRouter } from 'next/router'
+import AddFriendsHits from 'components/AddFriendsHits'
+import { Hits as YourFriendsHits } from 'components/YourFriendsHits'
 
 function zip(arrays) {
   return arrays[0].map((_, i) => {
@@ -91,90 +88,18 @@ const PopoutContainer = styled(Container)`
   width: 100%;
 `
 
-const AddFriend = styled.div`
-  width: 65px;
-  height: 65px;
-  background: ${(props) => props.theme.colors.white};
-  border-radius: 50%;
-  border: 3px solid #982649;
-  position: absolute;
-  top: 0;
-  right: 15px;
-  z-index: 50;
-`
-
-const Wrapper = styled.div`
-  position: relative;
-`
-
-const ImageContainer = styled(Center)`
-  height: 100%;
-  width: 100%;
-  position: relative;
-`
-
-const InvitationSend = styled.div`
-  display: block;
-  top: 0.05em;
-  left: 0.28em;
-  width: 15%;
-  height: 60%;
-  border: solid #982649;
-  border-width: 0 0.2em 0.2em 0;
-  transform: rotate(45deg);
-`
-
 const SearchBox = ({ currentRefinement, refine }) => {
   return (
     <Search
       placeholder={'Search players...'}
       type="search"
       value={currentRefinement}
-      onChange={(event) => refine(event.currentTarget.value)}
+      onChange={(e) => refine(e.currentTarget.value)}
     />
   )
 }
 
 const CustomSearchBox = connectSearchBox(SearchBox)
-
-const Hits = connectHits(({ hits, refine, currentRefinement }) => {
-  const [invitedFriends, setInvitedFriends] = useState([])
-  const router = useRouter()
-
-  const handleAddFriends = (friend) => {
-    !invitedFriends.includes(friend)
-      ? setInvitedFriends((current) => [...current, friend])
-      : setInvitedFriends((current) =>
-          current.filter((item) => item !== friend)
-        )
-    console.log(invitedFriends)
-  }
-
-  const items = hits.map((hit, idx) => (
-    <Wrapper key={idx}>
-      <UserColumn style={{ alignItems: 'center' }}>
-        <div onClick={() => router.push(`/profile/${hit.username}`)}>
-          <Avatar src={'/devonhenry_.svg'} />
-        </div>
-        <Username>{hit.usernanme}</Username>
-      </UserColumn>
-      <AddFriend>
-        <ImageContainer onClick={() => handleAddFriends(hit.usernanme)}>
-          {invitedFriends.includes(hit.usernanme) ? (
-            <InvitationSend />
-          ) : (
-            <Image src={'/cross.svg'} width={32} height={32} alt={'cross'} />
-          )}
-        </ImageContainer>
-      </AddFriend>
-    </Wrapper>
-  ))
-  return (
-    <Grid columns={'repeat(auto-fit, minmax(210px, 1fr))'} gap={'83px'}>
-      {items}
-    </Grid>
-  )
-})
 
 export default function Friends({ friends, invites, avatars }) {
   const [selected, setSelected] = useState('yourfriends')
@@ -187,7 +112,7 @@ export default function Friends({ friends, invites, avatars }) {
           <CustomSearchBox />
         </SearchContainer>
         <Grid columns={'repeat(auto-fit, minmax(210px, 1fr))'} gap={'83px'}>
-          <Hits />
+          <AddFriendsHits />
         </Grid>
       </div>
     )
@@ -201,7 +126,7 @@ export default function Friends({ friends, invites, avatars }) {
           <CustomSearchBox />
         </SearchContainer>
         <Grid columns={'repeat(auto-fit, minmax(210px, 1fr))'} gap={'83px'}>
-          <Hits />
+          <YourFriendsHits />
         </Grid>
       </div>
     )
